+++ Source/Services/MvcProject.Services.Data/DescriptionsService.cs
@@ -19,13 +19,13 @@
 
         public override Description GetByEncodedId(string id)
         {
-            var description = this.Repository.GetById(this.IdentifierProvider.DecodeIdToInt(id));
+            var description = this.Repository.GetById((int)this.IdentifierProvider.DecodeIdToInt(id));
             return description;
         }
 
         public override Description GetByEncodedIdFromNotDeleted(string id)
         {
-            var description = this.Repository.GetByIdFromNotDeleted(this.IdentifierProvider.DecodeIdToInt(id));
+            var description = this.Repository.GetByIdFromNotDeleted((int)this.IdentifierProvider.DecodeIdToInt(id));
             return description;
         }
     }
