+++ Source/Services/MvcProject.Services.Data/ProductsService.cs
@@ -19,13 +19,13 @@
 
         public override Product GetByEncodedId(string id)
         {
-            var product = this.Repository.GetById(this.IdentifierProvider.DecodeIdToInt(id));
+            var product = this.Repository.GetById((int)this.IdentifierProvider.DecodeIdToInt(id));
             return product;
         }
 
         public override Product GetByEncodedIdFromNotDeleted(string id)
         {
-            var product = this.Repository.GetByIdFromNotDeleted(this.IdentifierProvider.DecodeIdToInt(id));
+            var product = this.Repository.GetByIdFromNotDeleted((int)this.IdentifierProvider.DecodeIdToInt(id));
             return product;
         }
     }
