+++ Source/Services/MvcProject.Services.Data/TagsService.cs
@@ -19,13 +19,13 @@
 
         public override Tag GetByEncodedId(string id)
         {
-            var tag = this.Repository.GetById(this.IdentifierProvider.DecodeIdToInt(id));
+            var tag = this.Repository.GetById((int)this.IdentifierProvider.DecodeIdToInt(id));
             return tag;
         }
 
         public override Tag GetByEncodedIdFromNotDeleted(string id)
         {
-            var tag = this.Repository.GetByIdFromNotDeleted(this.IdentifierProvider.DecodeIdToInt(id));
+            var tag = this.Repository.GetByIdFromNotDeleted((int)this.IdentifierProvider.DecodeIdToInt(id));
             return tag;
         }
     }
